Update method Setup
Primarily setting up movement in Player and building the structure for
movement in missiles once a player is in game.  Finding an easy way to
differentiate power ups would be useful when parsing through power up
list in player.  Need a way to actually assign values to the powerUp
associated with each agent